+++ src/main/java/net/silentchaos512/loginar/setup/LsRecipeSerializers.java
@@ -1,12 +1,14 @@
 package net.silentchaos512.loginar.setup;
 
 import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.item.crafting.SimpleRecipeSerializer;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import net.silentchaos512.lib.crafting.recipe.ExtendedShapedRecipe;
 import net.silentchaos512.loginar.LoginarMod;
 import net.silentchaos512.loginar.crafting.recipe.UrnBaseRecipe;
+import net.silentchaos512.loginar.crafting.recipe.UrnModifcationRecipe;
 
 import java.util.function.Supplier;
 
@@ -16,11 +18,14 @@ public class LsRecipeSerializers {
     public static final RegistryObject<RecipeSerializer<?>> URN_BASE = register("urn_base",
             UrnBaseRecipe.Serializer::new
     );
+    public static final RegistryObject<SimpleRecipeSerializer<?>> URN_MODIFICATION = register("urn_modification",
+            () -> new SimpleRecipeSerializer<>(UrnModifcationRecipe::new)
+    );
     public static final RegistryObject<RecipeSerializer<?>> URN_UPGRADE = register("urn_upgrade",
             () -> ExtendedShapedRecipe.Serializer.basic(UrnBaseRecipe::new)
     );
 
-    private static RegistryObject<RecipeSerializer<?>> register(String name, Supplier<RecipeSerializer<?>> serializer) {
+    private static <T extends RecipeSerializer<?>> RegistryObject<T> register(String name, Supplier<T> serializer) {
         return REGISTER.register(name, serializer);
     }
 }