+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Urns can now acquire gem colors from Silent Gear materials (only if Silent Gear is installed)
+- Urns can now have their clay parts dyed any color by crafting them with dyes
+
 ## [0.1.0] - 2022-10-21
 - First alpha release!
 - Loginar mob (missing some AI, but it spawns and drops items)